solution v0.2 create components folder
